\name{googleChart}
\alias{googleChart}
\title{Generate a Google Chart object}
\usage{
googleChart(data, type = "Table", options = list())
}
\arguments{
  \item{data}{data frame to render as a Google chart. The
  mappings (xvar, yvar, etc.) are determined by column
  order and chart type.}

  \item{type}{chart type}

  \item{options}{chart options}
}
\description{
Generate a Google Chart object
}

