\name{renderGoogleChart}
\alias{renderGoogleChart}
\title{Render A Google Chart}
\usage{
renderGoogleChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
  \item{expr}{an expression that returns a data frame}

  \item{env}{The environment in which to evaluate
  \code{expr}}

  \item{quoted}{is expr a quoted expression? This is useful
  if you want to save an expression in a variable.}
}
\description{
Render A Google Chart
}

